Added data base

+++ src/main/java/com/gmail/aspoka1/HelloService.java
@@ -2,7 +2,7 @@ package com.gmail.aspoka1;
 
 class HelloService {
     static final String FALLBACK_NAME = "world";
-    static final Lang FALLBACK_LANG = new Lang(1L, "Hello", "eng");
+    static final Lang FALLBACK_LANG = new Lang(1, "Hello", "eng");
 
     LangRepository langRepository;
 
@@ -15,10 +15,10 @@ class HelloService {
     }
 
     String prepareGreeting(String name, String id){
-        long langId;
+        int langId;
         Lang lang;
         try {
-            langId = Long.parseLong(id);
+            langId = Integer.parseInt(id);
         } catch (NumberFormatException e) {
             langId = FALLBACK_LANG.getId();
         }
